Fixes/Refactoring to new APIs to accomodate HAT upstream merge
