+++ src/main/scala/hatdex/hat/api/service/DataDebitService.scala
@@ -18,7 +18,7 @@ trait DataDebitService extends BundleService {
                                (implicit session: Session, user: User): Try[ApiDataDebit] = {
     val dataDebitKey = UUID.randomUUID()
     val newDebit = DataDebitRow(dataDebitKey, LocalDateTime.now(), LocalDateTime.now(), debit.name,
-      debit.startDate, debit.endDate, debit.rolling, debit.sell, debit.price,
+      debit.startDate, debit.endDate, debit.rolling, debit.sellRent, debit.price,
       enabled = false, "owner", user.userId.toString,
       debit.bundleContextless.flatMap(bundle => bundle.id),
       None,
